+++ app/console/src/Commands/BuildCommand.php
@@ -48,7 +48,8 @@ class BuildCommand extends Command
         $filter = '/' . implode('|', $this->excludes) . '/i';
         $packages = [
             'greencheap/blog' => '*',
-            'greencheap/theme-one' => '*'
+            'greencheap/theme-one' => '*',
+            'greencheap/seo' => '*'
         ];
 
         $config = [];