+++ app/console/src/Commands/TranslationFetchCommand.php
@@ -43,7 +43,8 @@ class TranslationFetchCommand extends Command
         exec(sprintf("git clone %s %s", $repo, $tmp));
 
         // foreach resource:
-        $resources = ['system', 'blog', 'theme-one'];
+        //$resources = ['system', 'blog', 'theme-one'];
+        $resources = ['system'];
 
         // mv translation files to correct folder
         foreach ($resources as $resource) {